% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addMaxIter.R
\name{addMaxIter}
\alias{addMaxIter}
\title{Higher Order Function to Add a Maximum Iteration}
\usage{
addMaxIter(fun, maxIter)
}
\arguments{
\item{fun}{function to be wrapped}

\item{maxIter}{maximum number of iterations}
}
\description{
This function can be used to modify convergence criterion functions.
}

