% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\docType{methods}
\name{bootstrap}
\alias{bootstrap}
\alias{bootstrap,ANY,ANY,integerORnumeric-method}
\alias{bootstrap,ANY,missing,missing-method}
\alias{bootstrap,ANY,missing,numericORinteger-method}
\alias{bootstrap,rfh,rfh,NULL-method}
\title{Fit model on Bootstrap sample}
\usage{
bootstrap(modelA, modelB = modelA, B = NULL, ...)

\S4method{bootstrap}{ANY,missing,missing}(modelA, modelB = modelA, B = NULL,
  ...)

\S4method{bootstrap}{ANY,missing,numericORinteger}(modelA, modelB = modelA,
  B = NULL, ...)

\S4method{bootstrap}{ANY,ANY,integerORnumeric}(modelA, modelB, B,
  filter = NULL, ...)

\S4method{bootstrap}{rfh,rfh,`NULL`}(modelA, modelB = modelA, B = NULL, ...)
}
\arguments{
\item{modelA}{a fitted object}

\item{modelB}{a fitted object used to draw samples. In most cases this is
\code{modelA}. Alternatively it may be useful to use a non-robust model.}

\item{B}{the number of repetitions}

\item{...}{arguments passed down to methods}

\item{filter}{a vector indicating which elements in the fittedd object to
keep in each repetition.}
}
\description{
These methods help to repeatedly fit a \link{rfh} model on bootstrap samples.
}

