% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/countCalls.R
\docType{class}
\name{countCalls}
\alias{DataWithCount}
\alias{DataWithCount-class}
\alias{countCalls}
\title{Higher Order Function to Count the Number of Calls}
\usage{
countCalls(fun)
}
\description{
This function can be used to count the number of calls of a function. See \code{\link{fixedPoint}} for how it can be used.
}

