% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitrfh}
\alias{fitGenericModel}
\alias{fitrfh}
\alias{fitrsfh}
\alias{fitrstfh}
\alias{fitrtfh}
\title{Fitting Precedures}
\usage{
fitrfh(y, x, samplingVar, ...)

fitrsfh(y, x, samplingVar, W, x0Var = c(0.01, 1), ...)

fitrtfh(y, x, samplingVar, nTime, x0Var = c(0.01, 1, 1), ...)

fitrstfh(y, x, samplingVar, W, nTime, x0Var = c(0.01, 0.01, 1, 1), ...)

fitGenericModel(y, x, matVFun, fixedPointParam, k = 1.345, K = getK(k),
  x0Coef = NULL, x0Var = 1, x0Re = NULL, tol = 1e-06, maxIter = 100,
  maxIterParam = 1, maxIterRe = 100, convCrit = convCritRelative(tol))
}
\arguments{
\item{y}{(numeric) response vector}

\item{x}{([m|M]atrix) the design matrix}

\item{samplingVar}{(numeric) vector with sampling variances}

\item{...}{arguments passed to \code{fitGenericModel}}

\item{W}{(matrix) proximity matrix}

\item{x0Var}{(numeric) starting values for variance parameters}

\item{nTime}{(integer) number of time periods}

\item{matVFun}{(function) a function with one argument - the variance
parameters - constructing something like \link{matVFH}}

\item{fixedPointParam}{(function) a function with one argument. The vector of
model parameters. Returns a list of results of the next iteration in the
overall algorithm.}

\item{k}{(numeric) tuning constant}

\item{K}{(numeric) scaling constant}

\item{x0Coef}{(numeric) starting values for regression coefficients}

\item{x0Re}{(numeric) starting values for random effects}

\item{tol}{(numeric) numerical toloerance to be used during optimisation}

\item{maxIter}{(integer) the maximum number of iterations for model parameters.}

\item{maxIterParam}{(integer) the maximum number of iterations for each
parameter in each overall iteration}

\item{maxIterRe}{(integer) the maximum number of iterations for fitting the
random effects}

\item{convCrit}{(function) a function defining the stopping rule}

\item{psi}{(function) influence function}
}
\description{
Several fitting procedures. Not intended for interactive use.
}

