% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustObjectiveFunctions.R
\name{robustObjectiveDelta}
\alias{fixedPointNumericDelta}
\alias{fixedPointRobustBeta}
\alias{fixedPointRobustDelta}
\alias{fixedPointRobustDelta2}
\alias{fixedPointRobustRandomEffect}
\alias{robustObjectiveDelta}
\title{Fixed Point Functions}
\usage{
robustObjectiveDelta(y, x, beta, matVFun, psi, K, derivSelect)

fixedPointNumericDelta(y, x, beta, matVFun, psi, K, derivSelect, stepSize)

fixedPointRobustBeta(y, X, matV, psi)

fixedPointRobustDelta(y, X, beta, matVFun, psi, K, derivSelect = 1)

fixedPointRobustDelta2(y, X, beta, matVFun, psi, K, derivSelect)

fixedPointRobustRandomEffect(y, X, beta, matV, psi)
}
\arguments{
\item{y}{vector of response}

\item{beta}{beta coefficients to be used}

\item{matVFun}{a function with one argument constructing something similar to
\link{matVFH}: e.g. \code{. \%>\% matVFH(c(1, 1))}}

\item{psi}{influence function}

\item{K}{constant, see \link{getK}}

\item{derivSelect}{an indicator to select the derivative in matV. Position or
name}

\item{X}{design matrix}

\item{matV}{(list of functions) see \link{matVFH}}
}
\description{
This is an implementation of a robustified fixed point function to identify
beta coefficients in any mixed linear model.
}

