% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeXY.R
\name{makeXY}
\alias{makeXY}
\title{makeXY}
\usage{
makeXY(.formula, .data)
}
\arguments{
\item{.formula}{(formula)}

\item{.data}{(data.frame) data from which design matrix and response to
extract from}
}
\description{
Extract respone vector and design matrix from data with given
  formula.
}
\examples{
set.seed(1)
makeXY(y ~ I(x^2), data.frame(x = rnorm(10), y = rnorm(10)))
}
