% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matModels.R
\name{matVFH}
\alias{matVFH}
\alias{matVSFH}
\alias{matVSTFH}
\alias{matVTFH}
\title{Model Variance Covariance Structures}
\usage{
matVFH(.sigma2, .samplingVar)

matVSFH(.rho, .sigma2, .W, .samplingVar)

matVTFH(.rho, .sigma2, .nTime, .samplingVar)

matVSTFH(.rho, .sigma2, .W, .nTime, .samplingVar)
}
\arguments{
\item{.sigma2}{(numeric) a scalar. The variance parameter of the random
effects.}

\item{.samplingVar}{(numeric) the vector of sampling variances of the direct
estimator.}

\item{.rho}{(numeric) the correlation parameter for correlated random effects.}

\item{.W}{(matrix) proximity matrix.}

\item{.nTime}{(integer) the number of time periods.}
}
\description{
Functions to create the variance-covariance structure for the implemented
models. All these functions return a list with methods to compute different
parts of the variance structure.
}
\details{
\code{matVFH} constructs variance-covariance matrix for a FH-model.

\code{matVSFH} the spatial FH model.
}

