% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse}
\alias{mse}
\alias{mse.rfh}
\title{Compute the Mean Squared Error of an Estimator}
\usage{
mse(object, ...)

\method{mse}{rfh}(object, type = "pseudo", re = NULL, V = NULL, B = 100,
  ...)
}
\arguments{
\item{object}{(see methods) an object containing the estimation result, e.g.
\link{rfh}}

\item{...}{arguments passed to methods}

\item{type}{(character) the type of the MSE. Available are 'pseudo' and
'boot'}

\item{re}{(NULL | numeric) optionally provide the estimated random effects}

\item{V}{(NULL | list) optionally provide the variance using \link{variance}}

\item{B}{(numeric) number of bootstrap repetitions}
}
\description{
A generic function to compute the mean squared error of the predicted values
under the estimated model. This estimation can also be triggered using the
predict function.
}

