% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/newtonRaphson.R
\name{newtonRaphson}
\alias{newtonRaphson}
\title{Newton-Raphson Algorithm}
\usage{
newtonRaphson(funList, ...)
}
\arguments{
\item{funList}{the function to be evaluated in the algorithm}

\item{...}{arguments passed to \code{\link{fixedPoint}}}
}
\description{
Newton-Raphson finds zeroes of a function. The user can supply the function and its first derivative. Note that the Newton Raphson Algorithm is a special case of a fixed point algorithm thus it is implemented using \code{\link{fixedPoint}} and is only a convenience.
}
\examples{
\dontrun{
vignette("fixedPoint", "saeRobustTools")
}
}

