% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiFunctions.R
\name{psiOne}
\alias{psiOne}
\alias{getK}
\title{psiOne}
\usage{
psiOne(u, k = 1.345, deriv = FALSE)

getK(k)
}
\arguments{
\item{u}{standardized residuals}

\item{k}{tuning constant}

\item{deriv}{if \code{TRUE} returns the derivative}
}
\description{
\code{psiOne} is a Huber influence function. \code{getK}
  function to compute capital K -- used internally.
}
\examples{
set.seed(1)
u <- rnorm(10)
psiOne(u, k = 1.345, deriv = FALSE)
}
