% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psiFunctions.R
\name{psiOne}
\alias{psiOne}
\title{psiOne}
\usage{
psiOne(u, k = 1.345, deriv = FALSE)
}
\arguments{
\item{u}{standardized residuals}

\item{k}{tuning constant}

\item{deriv}{if \code{TRUE} returns the derivative}
}
\description{
influence-function
}

