% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rfh.R
\name{rfh}
\alias{rfh}
\title{Robust Fay Herriot}
\usage{
rfh(y, X, samplingVar, theta0 = c(rep(1, ncol(X)), 1))
}
\arguments{
\item{y}{the response}

\item{X}{the design matrix}

\item{samplingVar}{the vector of sampling variances}

\item{theta0}{starting values for}
}
\description{
Implementations of robust FH models
}

