% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/saveHistory.R
\docType{class}
\name{saveHistory}
\alias{DataWithHistory}
\alias{DataWithHistory-class}
\alias{saveHistory}
\title{Higher Order Function to Save the History of Results}
\usage{
saveHistory(fun)
}
\description{
This function can be used to save a history of results of a function. The history is stored as a matrix, so this works best if the return value of \code{fun} is numeric. See \code{\link{fixedPoint}} for how it can be used.
}

