% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustEstimationEquations.R
\name{score}
\alias{score}
\title{Compute values of robust score functions}
\usage{
score(object, filter, ...)
}
\arguments{
\item{object}{a fitted object}

\item{filter}{(character) a selection of values to be computed}

\item{...}{arguments passed to methods}
}
\description{
These values should be close to zero for a model fit.
}

