% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustObjectiveFunctions.R
\name{scoreRobustBeta}
\alias{scoreRobustBeta}
\title{Robust score function (ML) for beta}
\usage{
scoreRobustBeta(y, X, matV, psi)
}
\arguments{
\item{y}{vector of response}

\item{X}{design matrix}

\item{matV}{(list of functions) see \link{matVFH}}

\item{psi}{influence function}
}
\description{
Constructs a list of functions with \code{f} as the score and \code{f1} as
its derivative. Both are functions of the beta coefficients.
}

