% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\docType{methods}
\name{update,rfh-method}
\alias{update,rfh-method}
\alias{update,fitrfh-method}
\title{Update a fitted object}
\usage{
\S4method{update}{rfh}(object, formula, ..., where = parent.frame(2))

\S4method{update}{fitrfh}(object, ...)
}
\arguments{
\item{object}{(rfh) an object fitted by \link{rfh}}

\item{formula}{see \link{update.formula}}

\item{...}{arguments passed to \link{update.default}}

\item{where}{(environment) should not be specified by the user}
}
\description{
This is a method which can be used to update a \link{rfh} result object and
refit it. The fitted parameter values from the current object are used as
starting values, then \link{update.default} is called.
}
