% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{variance}
\alias{variance}
\alias{variance.rfh}
\alias{weights.rfh}
\title{Construct variance}
\usage{
variance(.object, ...)

\method{variance}{rfh}(.object, ...)

\method{weights}{rfh}(object, re = NULL, ...)
}
\arguments{
\item{.object, object}{an object}

\item{...}{arguments passed to method}

\item{re}{(numeric | NULL) if NULL then the random effects have to be
computet.}
}
\description{
A generic function to construct the different variance components of an
object.
}

