% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.R
\name{matU}
\alias{matA}
\alias{matAConst}
\alias{matB}
\alias{matBConst}
\alias{matTrace}
\alias{matU}
\alias{matVFH}
\alias{matW}
\title{Helper functions operating on matrices}
\usage{
matU(.V)

matTrace(x)

matVFH(.sigma2, .samplingVar)

matB(y, X, beta, u, matV, psi)

matBConst(y, X, beta, matV, psi)

matA(y, X, beta, matV, psi)

matAConst(y, X, matV, psi)

matW(y, X, beta, u, matV, psi)
}
\arguments{
\item{.V}{(Matrix) variance matrix}

\item{x}{([m|M]atrix) a matrix}

\item{.sigma2}{(numeric) a scalar. The variance parameter of the random
effects.}

\item{.samplingVar}{(numeric) the vector of sampling variances of the direct
estimator.}

\item{y}{(numeric) response}

\item{X}{(Matrix) design matrix}

\item{beta}{(numeric) vector of regression coefficients}

\item{u}{(numeric) vector of random effects}

\item{matV}{(list of functions) see \code{matVFH} as an example}

\item{psi}{(function) the influence function}
}
\description{
Helper functions operating on matrices
}
\details{
\code{matU} computes U. U is the matrix containing only the diagonal
  elements of V. See Sinha & Rao (2009): page 386 for the def of U.

\code{matTrace} computes the trace of a matrix.

\code{matVFH} constructs variance-covariance matrix for a FH-model.
  Returns a list with the matrix and its inverse.

\code{matB} computes the matrix B from the CCST paper which is used
  to compute the weights in the pseudo linearized representation of the
  REBLUP.

\code{matBConst} returns a function with one argument, u, to compute
  the matrix B. This function is used internally to compute B in the fixed
  point algorithm (see \link{fixedPointRobustRandomEffect}).

\code{matA} computes the matrix A from the CCST paper which is used
  to compute the weights in the pseudo linearized representation of the
  REBLUP.

\code{matAConst} returns a function with one argument, beta, to
  compute the matrix A. This function is used internally to compute A in the
  fixed point algorithm for beta (see \link{fixedPointRobustBeta}).

\code{matW} returns a matrix containing the weights as they are
  defined in the CCST pseudo linearization. \code{matW \%*\% y} is the
  REBLUP.
}

