% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustObjectiveFunctions.R
\name{fixedPointRobustBeta}
\alias{fixedPointRobustBeta}
\alias{fixedPointRobustRandomEffect}
\alias{fixedPointRobustVarianceFH}
\title{Fixed Point Functions}
\usage{
fixedPointRobustBeta(y, X, matV, psi)

fixedPointRobustVarianceFH(y, X, samplingVar, psi, K, beta)

fixedPointRobustRandomEffect(y, X, beta, matV, psi)
}
\arguments{
\item{y}{vector of response}

\item{X}{design matrix}

\item{matV}{(list of functions) see \link{matVFH}}

\item{psi}{influence function}

\item{samplingVar}{the sampling variance of the direct estimator.}

\item{K}{constant, see \link{getK}. \code{length(beta) == ncols(X)}}

\item{beta}{beta coefficients to be used.}
}
\description{
This is an implementation of a robustified fixed point function to identify
beta coefficients in any mixed linear model. The function is derived from the
pseudo linear form of robust mixed models in CCST (2015) and CCT (2011).
}

