% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustObjectiveFunctions.R
\name{fixedPointRobustBeta}
\alias{fixedPointNumericDelta}
\alias{fixedPointRobustBeta}
\alias{fixedPointRobustDelta}
\alias{fixedPointRobustDelta2}
\alias{fixedPointRobustRandomEffect}
\alias{robustObjectiveDelta}
\title{Fixed Point Functions}
\usage{
fixedPointRobustBeta(y, x, matV, psi)

robustObjectiveDelta(y, x, beta, matVFun, psi, K, derivSelect)

fixedPointNumericDelta(y, x, beta, matVFun, psi, K, derivSelect, stepSize,
  lowerBound)

fixedPointRobustDelta(y, x, beta, matVFun, psi, K, derivSelect = 1)

fixedPointRobustDelta2(y, x, beta, matVFun, psi, K, derivSelect)

fixedPointRobustRandomEffect(y, x, beta, matV, psi)
}
\arguments{
\item{y}{vector of response}

\item{x}{design matrix}

\item{matV}{(list of functions) see \link{matVFH}}

\item{psi}{influence function}

\item{beta}{beta coefficients to be used}

\item{matVFun}{a function with one argument constructing something similar to
\link{matVFH}: e.g. \code{. \%>\% matVFH(c(1, 1))}}

\item{K}{constant, see \link{getK}}

\item{derivSelect}{an indicator to select the derivative in matV. Position or
name}

\item{stepSize}{(numeric) size to be used in numeric derivative}

\item{lowerBound}{(numeric) a lower bound, such that \code{param - stepSize}
cannot be outside of the parameter space}
}
\description{
This is an implementation of robustified fixed point functions to identify
model parameters in any mixed linear model: regression coefficients, variance
parameters, and random effects
}

