% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustObjectiveFunctions.R
\name{fixedPointRobustBeta}
\alias{fixedPointRobustBeta}
\alias{fixedPointRobustDelta}
\alias{fixedPointRobustRandomEffect}
\title{Fixed Point Functions}
\usage{
fixedPointRobustBeta(y, X, matV, psi)

fixedPointRobustDelta(y, X, beta, matVFun, psi, K)

fixedPointRobustRandomEffect(y, X, beta, matV, psi)
}
\arguments{
\item{y}{vector of response}

\item{X}{design matrix}

\item{matV}{(list of functions) see \link{matVFH}}

\item{psi}{influence function}

\item{beta}{beta coefficients to be used}

\item{matVFun}{a function with one argument constructing something similar to
\link{matVFH}: e.g. \code{. \%>\% matVFH(c(1, 1))}}

\item{K}{constant, see \link{getK}}
}
\description{
This is an implementation of a robustified fixed point function to identify
beta coefficients in any mixed linear model. The function is derived from the
pseudo linear form of robust mixed models in CCST (2015) and CCT (2011).
}

