% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustObjectiveFunctions.R
\name{fixedPointRobustBeta}
\alias{fixedPointRobustBeta}
\alias{fixedPointRobustRandomEffect}
\alias{fixedPointRobustVarianceFH}
\title{Fixed Point Functions}
\usage{
fixedPointRobustBeta(y, X, V, Vinv = solve(V), psi, resid = NULL)

fixedPointRobustVarianceFH(y, X, samplingVar, psi, K, beta)

fixedPointRobustRandomEffect(y, X, beta, VuSqrtInv, VeSqrtInv, psi)
}
\arguments{
\item{y}{vector of response}

\item{X}{design matrix}

\item{V}{variance matrix}

\item{Vinv}{inverse of V (optional)}

\item{psi}{influence function}

\item{resid}{function to compute residuals (optional)}

\item{samplingVar}{the sampling variance of the direct estimator.}

\item{K}{constant, see \link{getK}. \code{length(beta) == ncols(X)}}

\item{beta}{beta coefficients to be used.}

\item{VuSqrtInv}{(Matrix) n times n matrix of the variance structure of the
random effects part to the power of -1/2}

\item{VeSqrtInv}{(Matrix) n times n matrix of the variance structure of the
model error part to the power of -1/2}
}
\description{
This is an implementation of a robustified fixed point function to identify beta coefficients in any mixed linear model. The function is derived from the pseudo linear form of robust mixed models in CCST (2015) and CCT (2011).
}

