% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse}
\alias{mse}
\alias{mse.fitrfh}
\title{Compute the Mean Squared Error of an Estimator}
\usage{
mse(object, ...)

\method{mse}{fitrfh}(object, type = "pseudo", predType = "reblupbc",
  B = 100, ...)
}
\arguments{
\item{object}{(see methods) an object containing the estimation result, e.g.
\link{rfh}}

\item{...}{arguments passed to methods}

\item{type}{(character) the type of the MSE. Available are 'pseudo' and
'boot'}

\item{predType}{(character) the type of prediction: \code{c("reblup",
"reblupbc")}}

\item{B}{(numeric) number of bootstrap repetitions}
}
\description{
A generic function to compute the mean squared error of the predicted values
under the estimated model. see \link{rfh} for an example.
}

