% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfh.R
\docType{methods}
\name{rfh}
\alias{fitrfh}
\alias{predict.rfh}
\alias{rfh}
\alias{rfh,formula,data.frame,character-method}
\title{Robust Fay Herriot Model}
\usage{
rfh(formula, data, samplingVar, ...)

\S4method{rfh}{formula,data.frame,character}(formula, data, samplingVar, ...)

fitrfh(y, X, samplingVar, x0 = c(rep(1, ncol(X)), 1), k = 1.345,
  tol = 1e-06, maxIter = 100)

\method{predict}{rfh}(object, type = "REBLUP", mse = "none", ...)
}
\arguments{
\item{formula}{(formula)}

\item{data}{(data.frame)}

\item{samplingVar}{(character)}

\item{...}{arguments passed to methods}

\item{y}{(numeric) response vector}

\item{X}{([m|M]atrix) the design matrix}

\item{x0}{(numeric) starting values}

\item{k}{(numeric) tuning constant}

\item{tol}{(numeric) numerical toloerance to be used during optimisation}

\item{object}{(rfh) an object of class rfh}

\item{type}{(character) one in \code{c("linear", "REBLUP")}}

\item{mse}{(character) which type of mse you want to compute for the
predictions. See \link{mse} for available options.}
}
\description{
Robust Fay Herriot Model
}

