% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R, R/rfh.R
\docType{methods}
\name{fitrfh}
\alias{fitrfh}
\alias{predict.fitrfh}
\alias{rfh}
\alias{rfh,formula,data.frame,character-method}
\title{Robust Fay Herriot Model}
\usage{
fitrfh(y, x, samplingVar, ...)

rfh(formula, data, samplingVar, ...)

\S4method{rfh}{formula,data.frame,character}(formula, data, samplingVar, ...)

\method{predict}{fitrfh}(object, type = "REBLUP", ...)
}
\arguments{
\item{y}{(numeric) response vector}

\item{x}{([m|M]atrix) the design matrix}

\item{samplingVar}{(character)}

\item{...}{arguments passed to methods}

\item{formula}{(formula)}

\item{data}{(data.frame)}

\item{object}{(rfh) an object of class rfh}

\item{type}{(character) one in \code{c("linear", "REBLUP")}}

\item{x0}{(numeric) starting values for variance parameters}

\item{k}{(numeric) tuning constant}

\item{tol}{(numeric) numerical toloerance to be used during optimisation}

\item{maxIter}{(integer) the maximum number of iterations}

\item{maxIterRe}{(integer) the maximum number of iterations for fitting the
random effects}
}
\description{
Robust Fay Herriot Model
}

