% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfh.R
\name{rfh}
\alias{predict.fitrfh}
\alias{rfh}
\alias{rfh,formula,data.frame,character,ANY-method}
\title{Robust Fay Herriot Model}
\usage{
rfh(formula, data, samplingVar, correlation = NULL, ...)

\S4method{rfh}{formula,data.frame,character,ANY}(formula, data, samplingVar, correlation, ...)

\method{predict}{fitrfh}(object, type = "reblup", ...)
}
\arguments{
\item{formula}{(formula)}

\item{data}{(data.frame)}

\item{samplingVar}{(character)}

\item{correlation}{an optional correlation structure, e.g. \link{corSAR1}.}

\item{...}{arguments passed \link{fitGenericModel}}

\item{object}{(rfh) an object of class rfh}

\item{type}{(character) one in \code{c("linear", "reblup")}}
}
\description{
Robust Fay Herriot Model
}

