% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfh.R
\docType{methods}
\name{rfh}
\alias{predict.fitrfh}
\alias{rfh}
\alias{rfh,formula,data.frame,character-method}
\title{Robust Fay Herriot Model}
\usage{
rfh(formula, data, samplingVar, ...)

\S4method{rfh}{formula,data.frame,character}(formula, data, samplingVar, ...)

\method{predict}{fitrfh}(object, type = "REBLUP", ...)
}
\arguments{
\item{formula}{(formula)}

\item{data}{(data.frame)}

\item{samplingVar}{(character)}

\item{...}{arguments passed \link{fitGenericModel}}

\item{object}{(rfh) an object of class rfh}

\item{type}{(character) one in \code{c("linear", "REBLUP")}}
}
\description{
Robust Fay Herriot Model
}

