% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rfh.R
\docType{methods}
\name{rfh}
\alias{predict.rfh}
\alias{rfh}
\alias{rfh,formula,data.frame,character-method}
\title{Robust Fay Herriot Model}
\usage{
rfh(formula, data, samplingVar, ...)

\S4method{rfh}{formula,data.frame,character}(formula, data, samplingVar, ...)

\method{predict}{rfh}(object, ...)
}
\arguments{
\item{formula}{(formula)}

\item{data}{(data.frame)}

\item{samplingVar}{(character)}

\item{...}{arguments passed to methods}

\item{object}{(rfh) an object of class rfh}
}
\description{
Robust Fay Herriot Model
}

