% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/robustObjectiveFunctions.R
\name{scoreRobustBeta}
\alias{scoreRobustBeta}
\title{Robust score function (ML) for beta}
\usage{
scoreRobustBeta(y, X, V, Vinv = solve(V), psi, resid = NULL)
}
\arguments{
\item{y}{vector of response}

\item{X}{design matrix}

\item{V}{variance matrix}

\item{Vinv}{inverse of V (optional)}

\item{psi}{influence function}

\item{resid}{function to compute residuals (optional)}
}
\description{
Constructs a list of functions with \code{f} as the score and \code{f1} as its derivative. Both are functions of the beta coefficients.
}

