% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{variance}
\alias{variance}
\alias{variance.fitrfh}
\alias{variance.fitrsfh}
\alias{variance.fitrstfh}
\alias{variance.fitrtfh}
\alias{weights.fitrfh}
\title{Construct variance}
\usage{
variance(.object, ...)

\method{variance}{fitrfh}(.object, ...)

\method{variance}{fitrsfh}(.object, ...)

\method{variance}{fitrtfh}(.object, ...)

\method{variance}{fitrstfh}(.object, ...)

\method{weights}{fitrfh}(object, c = 1, ...)
}
\arguments{
\item{.object, object}{an object}

\item{...}{arguments passed to method}

\item{c}{(numeric) scalar}
}
\description{
A generic function to construct the different variance components of an
object.
}

