% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{variance}
\alias{variance}
\alias{variance.fitrfh}
\alias{variance.fitrsfh}
\alias{weights.fitrfh}
\title{Construct variance}
\usage{
variance(.object, ...)

\method{variance}{fitrfh}(.object, ...)

\method{variance}{fitrsfh}(.object, ...)

\method{weights}{fitrfh}(object, ...)
}
\arguments{
\item{.object, object}{an object}

\item{...}{arguments passed to method}
}
\description{
A generic function to construct the different variance components of an
object.
}

