% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.R
\name{matU}
\alias{matA}
\alias{matAConst}
\alias{matB}
\alias{matBConst}
\alias{matTrace}
\alias{matU}
\alias{matW}
\title{Matrix constructor functions}
\usage{
matU(.V)

matTrace(x)

matB(y, X, beta, u, matV, psi)

matBConst(y, X, beta, matV, psi)

matA(y, X, beta, matV, psi)

matAConst(y, X, matV, psi)

matW(y, X, beta, u, matV, psi)
}
\arguments{
\item{.V}{(Matrix) variance matrix}

\item{x}{([m|M]atrix) a matrix}

\item{y}{(numeric) response}

\item{X}{(Matrix) design matrix}

\item{beta}{(numeric) vector of regression coefficients}

\item{u}{(numeric) vector of random effects}

\item{matV}{(list of functions) see \code{matVFH} for an example}

\item{psi}{(function) the influence function}
}
\description{
Matrix constructor functions
}
\details{
\code{matU} computes U. U is the matrix containing only the diagonal
  elements of V. This function returns a list of functions which can be
  called to compute specific transformations of U.

\code{matTrace} computes the trace of a matrix.

\code{matB} computes the matrix B which is used to compute the
  weights in the pseudo linearised representation of the REBLUP.

\code{matBConst} returns a function with one argument, u, to compute
  the matrix B. This function is used internally to compute B in the fixed
  point algorithm (see \link{fixedPointRobustRandomEffect}).

\code{matA} computes the matrix A which is used to compute the
  weights in the pseudo linearized representation of the REBLUP.

\code{matAConst} returns a function with one argument, beta, to
  compute the matrix A. This function is used internally to compute A in the
  fixed point algorithm for beta (see \link{fixedPointRobustBeta}).

\code{matW} returns a matrix containing the weights as they are
  defined for the pseudo linear form, such that \code{matW \%*\% y} is the
  REBLUP.
}

