% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.R
\name{matU}
\alias{matTrace}
\alias{matU}
\alias{matVFH}
\title{Helper functions operating on matrices and matrix constructors}
\usage{
matU(V)

matVFH(sigma2, samplingVar)

matTrace(x)
}
\arguments{
\item{V}{variance matrix}

\item{sigma2}{a scalar. The variance parameter of the random effects.}

\item{samplingVar}{the vector of sampling variances of the direct estimator.}

\item{x}{a matrix}
}
\description{
Helper functions operating on matrices and matrix constructors
}
\details{
\code{matU} computes U. U is the matrix containing only the diagonal elements of V. See Sinha & Rao (2009): page 386 for the def of U.

\code{matVFH} constructs variance-covariance matrix for a FH-model. Returns a list with the matrix and its inverse.

\code{matTrace} computes the trace of a matrix.
}

